refactor: fx-core exports clean up
