+++ packages/cli/src/commands/models/create.ts
@@ -14,10 +14,8 @@ import {
   CliQuestionName,
   CreateProjectInputs,
   CreateProjectOptions,
-  FeatureFlags,
   MeArchitectureOptions,
   QuestionNames,
-  featureFlagManager,
 } from "@microsoft/teamsfx-core";
 import chalk from "chalk";
 import { assign } from "lodash";
