+++ packages/cli/src/commands/models/teamsapp/validate.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 import { CLICommand, CLICommandOption, TeamsAppInputs, err } from "@microsoft/teamsfx-api";
+import { FeatureFlags, featureFlagManager } from "@microsoft/teamsfx-core";
 import { getFxCore } from "../../../activate";
 import { commands } from "../../../resource";
 import { TelemetryEvent } from "../../../telemetry/cliTelemetryEvents";
@@ -15,7 +16,6 @@ import {
   ValidateMethodOption,
 } from "../../common";
 import { validateArgumentConflict } from "./update";
-import { isAsyncAppValidationEnabled } from "../../../../../fx-core/build";
 
 export const teamsappValidateCommand: CLICommand = {
   name: "validate",
@@ -48,7 +48,7 @@ function getOptions(): CLICommandOption[] {
     ProjectFolderOption,
   ];
 
-  if (isAsyncAppValidationEnabled()) {
+  if (featureFlagManager.getBooleanValue(FeatureFlags.AsyncAppValidation)) {
     options.push(ValidateMethodOption);
   }
