+++ packages/fx-core/src/common/telemetry.ts
@@ -3,7 +3,6 @@
 
 import { FxError, SystemError } from "@microsoft/teamsfx-api";
 import { assign } from "lodash";
-import { TelemetryConstants } from "../component/constants";
 import { TOOLS, globalVars } from "./globalVars";
 import { ProjectTypeResult } from "./projectTypeChecker";
 import { maskSecret } from "./stringUtils";
@@ -72,6 +71,38 @@ export enum TelemetryProperty {
   HasAzureOpenAIDeploymentName = "has-azure-openai-deployment-name",
   HasOpenAIKey = "has-openai-key",
 }
+
+export const TelemetryConstants = {
+  eventPrefix: "-start",
+  properties: {
+    component: "component",
+    appId: "appid",
+    tenantId: "tenant-id",
+    success: "success",
+    errorCode: "error-code",
+    errorType: "error-type",
+    errorMessage: "err-message", // change the error message property key
+    errorStack: "err-stack", // change the error stack property key
+    timeCost: "time-cost",
+    errorName: "error-name", // need classify, keep error name as a separate property for telemetry analysis, error name should has limited set of values
+    innerError: "inner-error", // need classify, JSON serialized raw inner error that is caused by internal error or external call error
+    errorCat: "error-cat", // need classify, error category
+    errorCat1: "error-cat1", // need classify, error category level 1
+    errorCat2: "error-cat2", // need classify, error category level 2
+    errorCat3: "error-cat3", // need classify, error category level 3
+    errorStage: "error-stage", // need classify
+    errorComponent: "error-component", // need classify
+    errorMethod: "error-method", // need classify
+    errorSource: "error-source", // need classify
+    errorInnerCode: "error-inner-code", // need classify
+  },
+  values: {
+    yes: "yes",
+    no: "no",
+    userError: "user",
+    systemError: "system",
+  },
+};
 
 export enum TelemetryEvent {
   Scaffold = "scaffold",
