+++ packages/fx-core/src/component/coordinator/index.ts
@@ -23,7 +23,7 @@ import { EOL } from "os";
 import * as path from "path";
 import * as uuid from "uuid";
 import * as xml2js from "xml2js";
-import { getResourceGroupInPortal } from "../../common/constants";
+import { AppStudioScopes, getResourceGroupInPortal } from "../../common/constants";
 import { isNewGeneratorEnabled } from "../../common/featureFlags";
 import { getLocalizedString } from "../../common/localizeUtils";
 import { convertToAlphanumericOnly } from "../../common/stringUtils";
@@ -57,7 +57,7 @@ import { deployUtils } from "../deployUtils";
 import { developerPortalScaffoldUtils } from "../developerPortalScaffoldUtils";
 import { DriverContext } from "../driver/interface/commonArgs";
 import { updateTeamsAppV3ForPublish } from "../driver/teamsApp/appStudio";
-import { AppStudioScopes, Constants } from "../driver/teamsApp/constants";
+import { Constants } from "../driver/teamsApp/constants";
 import { CopilotPluginGenerator } from "../generator/copilotPlugin/generator";
 import { Generator } from "../generator/generator";
 import { Generators } from "../generator/generatorProvider";
