+++ packages/fx-core/src/component/driver/teamsApp/appStudio.ts
@@ -26,6 +26,7 @@ import { basename, extname } from "path";
 import { Container } from "typedi";
 import * as util from "util";
 import isUUID from "validator/lib/isUUID";
+import { AppStudioScopes } from "../../../common/constants";
 import { getDefaultString, getLocalizedString } from "../../../common/localizeUtils";
 import { FileNotFoundError, UserCancelError } from "../../../error/common";
 import { QuestionNames } from "../../../question/constants";
@@ -33,7 +34,7 @@ import { envUtil } from "../../utils/envUtil";
 import { DriverContext } from "../interface/commonArgs";
 import { AppStudioClient } from "./clients/appStudioClient";
 import { ConfigureTeamsAppDriver, actionName as configureTeamsAppActionName } from "./configure";
-import { AppStudioScopes, Constants, supportedLanguageCodes } from "./constants";
+import { Constants, supportedLanguageCodes } from "./constants";
 import {
   CreateAppPackageDriver,
   actionName as createAppPackageActionName,
