+++ packages/fx-core/src/component/driver/teamsApp/clients/appStudioClient.ts
@@ -5,48 +5,43 @@
 /**
  * @author yuqizhou77 <86260893+yuqizhou77@users.noreply.github.com>
  */
+import { LogProvider, SystemError } from "@microsoft/teamsfx-api";
 import axios, { AxiosInstance } from "axios";
-import { SystemError, LogProvider } from "@microsoft/teamsfx-api";
-import { AppDefinition } from "../../../driver/teamsApp/interfaces/appdefinitions/appDefinition";
-import { AppUser } from "../../../driver/teamsApp/interfaces/appdefinitions/appUser";
-import { AppStudioError } from ".././errors";
-import { IPublishingAppDenition } from "../interfaces/appdefinitions/IPublishingAppDefinition";
-import { AppStudioResultFactory } from ".././results";
-import {
-  Constants,
-  ErrorMessages,
-  APP_STUDIO_API_NAMES,
-  getAppStudioEndpoint,
-} from ".././constants";
-import { RetryHandler } from "../utils/utils";
-import { HelpLinks } from "../../../../common/constants";
+import { HelpLinks, getAppStudioEndpoint } from "../../../../common/constants";
+import { setErrorContext } from "../../../../common/globalVars";
 import { getLocalizedString } from "../../../../common/localizeUtils";
 import {
   Component,
-  sendTelemetryErrorEvent,
-  sendTelemetryEvent,
   TelemetryEvent,
   TelemetryProperty,
+  sendTelemetryErrorEvent,
+  sendTelemetryEvent,
 } from "../../../../common/telemetry";
 import { waitSeconds } from "../../../../common/utils";
-import { IValidationResult } from "../../../driver/teamsApp/interfaces/appdefinitions/IValidationResult";
-import { HttpStatusCode } from "../../../constant/commonConstant";
-import { manifestUtils } from "../utils/ManifestUtils";
-import { setErrorContext } from "../../../../common/globalVars";
+import { WrappedAxiosClient } from "../../../../common/wrappedAxiosClient";
 import {
   CheckSideloadingPermissionFailedError,
   DeveloperPortalAPIFailedError,
 } from "../../../../error/teamsApp";
+import { HttpStatusCode } from "../../../constant/commonConstant";
+import { IValidationResult } from "../../../driver/teamsApp/interfaces/appdefinitions/IValidationResult";
+import { AppDefinition } from "../../../driver/teamsApp/interfaces/appdefinitions/appDefinition";
+import { AppUser } from "../../../driver/teamsApp/interfaces/appdefinitions/appUser";
+import { APP_STUDIO_API_NAMES, Constants, ErrorMessages } from ".././constants";
+import { AppStudioError } from ".././errors";
+import { AppStudioResultFactory } from ".././results";
 import {
   ApiSecretRegistration,
   ApiSecretRegistrationUpdate,
 } from "../interfaces/ApiSecretRegistration";
-import { WrappedAxiosClient } from "../../../../common/wrappedAxiosClient";
+import { AsyncAppValidationDetailsResponse } from "../interfaces/AsyncAppValidationDetailsResponse";
 import { AsyncAppValidationResponse } from "../interfaces/AsyncAppValidationResponse";
 import { AsyncAppValidationResultsResponse } from "../interfaces/AsyncAppValidationResultsResponse";
-import { AsyncAppValidationDetailsResponse } from "../interfaces/AsyncAppValidationDetailsResponse";
-import { OauthRegistration } from "../interfaces/OauthRegistration";
 import { OauthConfigurationId } from "../interfaces/OauthConfigurationId";
+import { OauthRegistration } from "../interfaces/OauthRegistration";
+import { IPublishingAppDenition } from "../interfaces/appdefinitions/IPublishingAppDefinition";
+import { manifestUtils } from "../utils/ManifestUtils";
+import { RetryHandler } from "../utils/utils";
 
 // eslint-disable-next-line @typescript-eslint/no-namespace
 export namespace AppStudioClient {
