+++ packages/fx-core/src/component/driver/teamsApp/validateAppPackage.ts
@@ -23,6 +23,7 @@ import { merge } from "lodash";
 import { EOL } from "os";
 import * as path from "path";
 import { Service } from "typedi";
+import { AppStudioScopes } from "../../../common/constants";
 import { getDefaultString, getLocalizedString } from "../../../common/localizeUtils";
 import { FileNotFoundError, InvalidActionInputError } from "../../../error/common";
 import { SummaryConstant } from "../../configManager/constant";
@@ -32,7 +33,7 @@ import { ExecutionResult, StepDriver } from "../interface/stepDriver";
 import { addStartAndEndTelemetry } from "../middleware/addStartAndEndTelemetry";
 import { WrapDriverContext } from "../util/wrapUtil";
 import { AppStudioClient } from "./clients/appStudioClient";
-import { AppStudioScopes, Constants } from "./constants";
+import { Constants } from "./constants";
 import { AppStudioError } from "./errors";
 import { ValidateAppPackageArgs } from "./interfaces/ValidateAppPackageArgs";
 import { AppStudioResultFactory } from "./results";
