+++ packages/fx-core/src/component/driver/teamsApp/validateTestCases.ts
@@ -1,45 +1,41 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { EOL } from "os";
+import { hooks } from "@feathersjs/hooks/lib";
 import {
-  Result,
+  Colors,
   FxError,
-  ok,
-  err,
-  TeamsAppManifest,
   ManifestUtil,
   Platform,
-  Colors,
+  Result,
+  TeamsAppManifest,
+  err,
+  ok,
 } from "@microsoft/teamsfx-api";
-import { hooks } from "@feathersjs/hooks/lib";
-import { Service } from "typedi";
-import fs from "fs-extra";
-import * as path from "path";
-import { merge } from "lodash";
-import { StepDriver, ExecutionResult } from "../interface/stepDriver";
-import { DriverContext } from "../interface/commonArgs";
-import { WrapDriverContext } from "../util/wrapUtil";
-import { ValidateWithTestCasesArgs } from "./interfaces/ValidateWithTestCasesArgs";
-import { addStartAndEndTelemetry } from "../middleware/addStartAndEndTelemetry";
-import { AppStudioClient } from "./clients/appStudioClient";
-import { getLocalizedString } from "../../../common/localizeUtils";
 import AdmZip from "adm-zip";
-import {
-  Constants,
-  getAppStudioEndpoint,
-  CEHCK_VALIDATION_RESULTS_INTERVAL_SECONDS,
-  AppStudioScopes,
-} from "./constants";
-import { metadataUtil } from "../../utils/metadataUtil";
+import fs from "fs-extra";
+import { merge } from "lodash";
+import { EOL } from "os";
+import * as path from "path";
+import { Service } from "typedi";
+import { AppStudioScopes, getAppStudioEndpoint } from "../../../common/constants";
+import { getLocalizedString } from "../../../common/localizeUtils";
+import { waitSeconds } from "../../../common/utils";
 import { FileNotFoundError, InvalidActionInputError } from "../../../error/common";
+import { SummaryConstant } from "../../configManager/constant";
+import { metadataUtil } from "../../utils/metadataUtil";
+import { DriverContext } from "../interface/commonArgs";
+import { ExecutionResult, StepDriver } from "../interface/stepDriver";
+import { addStartAndEndTelemetry } from "../middleware/addStartAndEndTelemetry";
+import { WrapDriverContext } from "../util/wrapUtil";
+import { AppStudioClient } from "./clients/appStudioClient";
+import { CEHCK_VALIDATION_RESULTS_INTERVAL_SECONDS, Constants } from "./constants";
 import {
   AsyncAppValidationResponse,
   AsyncAppValidationStatus,
 } from "./interfaces/AsyncAppValidationResponse";
 import { AsyncAppValidationResultsResponse } from "./interfaces/AsyncAppValidationResultsResponse";
-import { SummaryConstant } from "../../configManager/constant";
-import { waitSeconds } from "../../../common/utils";
+import { ValidateWithTestCasesArgs } from "./interfaces/ValidateWithTestCasesArgs";
 
 const actionName = "teamsApp/validateWithTestCases";
