+++ packages/fx-core/src/component/feature/collaboration.ts
@@ -21,8 +21,9 @@ import { AadAppClient } from "../driver/aad/utility/aadAppClient";
 import { permissionsKeys } from "../driver/aad/utility/constants";
 import { addStartAndEndTelemetry } from "../driver/middleware/addStartAndEndTelemetry";
 import { AppStudioClient } from "../driver/teamsApp/clients/appStudioClient";
-import { AppStudioScopes, Constants } from "../driver/teamsApp/constants";
+import { Constants } from "../driver/teamsApp/constants";
 import { AppUser } from "../driver/teamsApp/interfaces/appdefinitions/appUser";
+import { AppStudioScopes } from "../../common/constants";
 
 const EventName = {
   grantPermission: "grant-permission",
