+++ packages/fx-core/src/component/provisionUtils.ts
@@ -13,7 +13,7 @@ import {
   SubscriptionInfo,
   UserError,
 } from "@microsoft/teamsfx-api";
-import { HelpLinks } from "../common/constants";
+import { AppStudioScopes, HelpLinks } from "../common/constants";
 import { getLocalizedString } from "../common/localizeUtils";
 import { TelemetryEvent, TelemetryProperty } from "../common/telemetry";
 import { getHashedEnv } from "../common/stringUtils";
@@ -32,7 +32,6 @@ import {
 } from "../error/m365";
 import { SolutionTelemetryProperty } from "./constants";
 import { DriverContext } from "./driver/interface/commonArgs";
-import { AppStudioScopes } from "./driver/teamsApp/constants";
 import { resourceGroupHelper, ResourceGroupInfo } from "./utils/ResourceGroupHelper";
 export interface M365TenantRes {
   tenantIdInToken: string;
