+++ packages/fx-core/src/component/resource/botService/botRegistration/botFrameworkRegistration.ts
@@ -5,10 +5,10 @@
  * @author Qianhao Dong <qidon@microsoft.com>
  */
 import { FxError, LogProvider, M365TokenProvider, Result, err, ok } from "@microsoft/teamsfx-api";
-import { AppStudioScopes } from "../../../driver/teamsApp/constants";
 import { AppStudioClient } from "../appStudio/appStudioClient";
 import { IBotRegistration } from "../appStudio/interfaces/IBotRegistration";
 import { Utils } from "./utils";
+import { AppStudioScopes } from "../../../../common/constants";
 
 export async function createOrUpdateBotRegistration(
   m365TokenProvider: M365TokenProvider,
