+++ packages/vscode-extension/src/debug/taskTerminal/launchTeamsClientTerminal.ts
@@ -1,28 +1,30 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import * as cp from "child_process";
-import * as vscode from "vscode";
-import * as util from "util";
-import * as globalVariables from "../../globalVariables";
 import { err, FxError, ok, Result, UserError, Void } from "@microsoft/teamsfx-api";
-import { BaseTaskTerminal } from "./baseTaskTerminal";
-import { Correlator } from "@microsoft/teamsfx-core";
-import { localTelemetryReporter, maskValue } from "../localTelemetryReporter";
-import { getLocalDebugSession } from "../commonUtils";
+import {
+  CoreQuestionNames,
+  Correlator,
+  environmentNameManager,
+  HubOptions,
+} from "@microsoft/teamsfx-core";
+import * as cp from "child_process";
+import * as util from "util";
+import * as vscode from "vscode";
 import VsCodeLogInstance from "../../commonlib/log";
+import { ExtensionErrors, ExtensionSource } from "../../error";
+import * as globalVariables from "../../globalVariables";
+import { core, getSystemInputs } from "../../handlers";
 import { TelemetryEvent, TelemetryProperty } from "../../telemetry/extTelemetryEvents";
-import { SolutionSource } from "@microsoft/teamsfx-core";
-import { ExtensionErrors } from "../../error";
 import { getDefaultString, localize } from "../../utils/localizeUtils";
+import { getLocalDebugSession } from "../commonUtils";
 import {
   launchingTeamsClientDisplayMessages,
   openTerminalDisplayMessage,
   openTerminalMessage,
 } from "../constants";
-import { core, getSystemInputs } from "../../handlers";
-import { CoreQuestionNames, environmentNameManager } from "@microsoft/teamsfx-core";
-import { HubOptions } from "@microsoft/teamsfx-core";
+import { localTelemetryReporter, maskValue } from "../localTelemetryReporter";
+import { BaseTaskTerminal } from "./baseTaskTerminal";
 
 interface LaunchTeamsClientArgs {
   env?: string;
@@ -110,7 +112,7 @@ export class LaunchTeamsClientTerminal extends BaseTaskTerminal {
         resolve(
           err(
             new UserError(
-              SolutionSource,
+              ExtensionSource,
               ExtensionErrors.LaunchTeamsWebClientError,
               `${getDefaultString("teamstoolkit.localDebug.launchTeamsWebClientError")} ${
                 error?.message ?? ""
@@ -130,7 +132,7 @@ export class LaunchTeamsClientTerminal extends BaseTaskTerminal {
           resolve(
             err(
               new UserError(
-                SolutionSource,
+                ExtensionSource,
                 ExtensionErrors.LaunchTeamsWebClientError,
                 util.format(
                   getDefaultString("teamstoolkit.localDebug.launchTeamsWebClientStoppedError"),
